Add simple I/O tests to test/niosd_io_test.c
This patch adds some I/O tests to the niosd_io_test.c test program.
Currently, these test I/Os are just reads of sector 0, however, the
reads are concurrent.

The patch also includes numerous cleanups and fixes in
src/niosd_io.c.  These were to be expected since before this test,
most of src/niosd_io.c had not been run at all.

Many more test cases are needed since it's likely that src/niosd_io.c
still has numerous problems.

+++ src/include/log.h
@@ -50,8 +50,8 @@ ll_to_string(enum log_level ll)
     {                                                                   \
         struct timespec ts;                                             \
         clock_gettime(CLOCK_MONOTONIC_RAW, &ts);                        \
-        fprintf(stderr, "<%ld.%05hu:%s:%lx:%s@%d> " message "\n",       \
-                ts.tv_sec, (unsigned short)ts.tv_nsec,                  \
+        fprintf(stderr, "<%ld.%lu:%s:%lx:%s@%d> " message "\n",       \
+                ts.tv_sec, ts.tv_nsec,                  \
                 ll_to_string(lvl), thread_id_get(), __func__,           \
                 __LINE__, ##__VA_ARGS__);                               \
         if (lvl == LL_FATAL)                                            \
